Issue 1780: UserRepo delete needs to handle case when active filter is NULL.
This problem is exposed by the unit tests.

+++ src/main/java/org/tdl/vireo/model/repo/impl/UserRepoImpl.java
@@ -2,9 +2,10 @@ package org.tdl.vireo.model.repo.impl;
 
 import static edu.tamu.weaver.response.ApiStatus.SUCCESS;
 
+import edu.tamu.weaver.data.model.repo.impl.AbstractWeaverRepoImpl;
+import edu.tamu.weaver.response.ApiResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
-import org.tdl.vireo.model.NamedSearchFilterGroup;
 import org.tdl.vireo.model.Role;
 import org.tdl.vireo.model.User;
 import org.tdl.vireo.model.repo.NamedSearchFilterGroupRepo;
@@ -12,9 +13,6 @@ import org.tdl.vireo.model.repo.UserRepo;
 import org.tdl.vireo.model.repo.custom.UserRepoCustom;
 import org.tdl.vireo.service.DefaultFiltersService;
 import org.tdl.vireo.service.DefaultSubmissionListColumnService;
-
-import edu.tamu.weaver.data.model.repo.impl.AbstractWeaverRepoImpl;
-import edu.tamu.weaver.response.ApiResponse;
 
 public class UserRepoImpl extends AbstractWeaverRepoImpl<User, UserRepo> implements UserRepoCustom {
 
@@ -73,7 +71,10 @@ public class UserRepoImpl extends AbstractWeaverRepoImpl<User, UserRepo> impleme
 
     @Override
     public void delete(User user) {
-        namedSearchFilterGroupRepo.delete(user.getActiveFilter());
+        if (user.getActiveFilter() != null) {
+            namedSearchFilterGroupRepo.delete(user.getActiveFilter());
+        }
+
         userRepo.deleteById(user.getId());
         simpMessagingTemplate.convertAndSend("/channel/user/delete", new ApiResponse(SUCCESS));
     }
